RavenDB-18201 - use one context in AbstractCountersHandlerProcessorForGetCounters - simply the abstraction
