+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/Counters/ShardedCountersHandlerProcessorForGetCounters.cs
@@ -4,7 +4,6 @@ using Microsoft.Extensions.Primitives;
 using Raven.Client.Documents.Operations.Counters;
 using Raven.Server.Documents.Handlers.Processors.Counters;
 using Raven.Server.ServerWide.Context;
-using Sparrow.Json;
 
 namespace Raven.Server.Documents.Sharding.Handlers.Processors.Counters
 {
@@ -14,15 +13,11 @@ namespace Raven.Server.Documents.Sharding.Handlers.Processors.Counters
         {
         }
 
-        protected override async ValueTask<CountersDetail> GetCountersAsync(string docId, StringValues counters, bool full)
+        protected override async ValueTask<CountersDetail> GetCountersAsync(TransactionOperationContext context, string docId, StringValues counters, bool full)
         {
             var op = new GetCountersOperation.GetCounterValuesCommand(docId, counters, full);
 
-            int shardNumber;
-            using (RequestHandler.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            {
-                shardNumber = RequestHandler.DatabaseContext.GetShardNumber(context, docId);
-            }
+            var shardNumber = RequestHandler.DatabaseContext.GetShardNumber(context, docId);
 
             using (var token = RequestHandler.CreateOperationToken())
             {
